+++ RELEASES.md
@@ -6,6 +6,7 @@ Ladybug Test Tool release notes
 Upcoming
 --------
 
+- Make ReportRunner public (make it callable for LadybugPipe)
 - Bugfix broken searching/filtering on EndTime
 - Replace Download all with Download table and Download tree
 - Add Open all